Create API29MediaContentMaanger, API21MediaContentManager class
